requirements.txt
praw==7.7.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
rich==13.7.1
openai==1.34.0
scikit-learn==1.5.1
textblob==0.15.3
nltk==3.8.1